test(cloudflare): split provider tests
